moved dummy test data to own class
