+++ Tabloid-Fullstack.Tests/PostTests/PostRepositoryTests.cs
@@ -7,7 +7,7 @@ using Xunit;
 // We are able to do these tests because we "spoof" the ApplicationDbContext using Entity Framework and the EFTextFixture Class
 // For each test, we create a new database in memory. Add some seed data, run our test, then delete Db.
 // This way we have total control of our test environment.
-namespace Tabloid_Fullstack.Tests
+namespace Tabloid_Fullstack.Tests.PostTests
 {
     // Inherits the EFTestFixture so it has access to our fake database 
     public class PostRepositoryTests : EFTestFixture
@@ -15,8 +15,8 @@ namespace Tabloid_Fullstack.Tests
 
         public PostRepositoryTests()
         {
-            // When constructed, runs dummy data
-            AddSampleData();
+            // When constructed, generate dummy data, passing in EFTestFixture database
+            PostDummyData.GenerateData(_context);
         }
 
         [Fact]
@@ -121,121 +121,6 @@ namespace Tabloid_Fullstack.Tests
 
             // The post we just added should have the default URL string.
             Assert.True(justAddedPost.ImageLocation == "http://lorempixel.com/920/360/");
-        }
-
-
-        // Dummy data
-        private void AddSampleData()
-        {
-            var userType1 = new UserType()
-            {
-                Name = "admin"
-            };
-
-            var userType2 = new UserType()
-            {
-                Name = "owner"
-            };
-
-            _context.Add(userType1);
-            _context.Add(userType2);
-            _context.SaveChanges();
-
-            var user1 = new UserProfile()
-            {
-                FirebaseUserId = "TEST_FIREBASE_UID_1",
-                DisplayName = "Dean",
-                FirstName = "Michael",
-                LastName = "Melchiondo",
-                Email = "dean@ween.com",
-                CreateDateTime = DateTime.Now - TimeSpan.FromDays(365),
-                ImageLocation = null,
-                UserTypeId =  1
-            };
-
-            var user2 = new UserProfile()
-            {
-                FirebaseUserId = "TEST_FIREBASE_UID_2",
-                DisplayName = "Gene",
-                FirstName = "Aaron",
-                LastName = "Freeman",
-                Email = "gene@ween.com",
-                CreateDateTime = DateTime.Now - TimeSpan.FromDays(365),
-                ImageLocation = null,
-                UserTypeId = 2
-            };
-
-            var user3 = new UserProfile()
-            {
-                FirebaseUserId = "TEST_FIREBASE_UID_3",
-                DisplayName = "bestDrummer",
-                FirstName = "Claude",
-                LastName = "Coleman",
-                Email = "claude@ween.com",
-                CreateDateTime = DateTime.Now - TimeSpan.FromDays(365),
-                ImageLocation = null,
-                UserTypeId = 2
-            };
-
-            _context.Add(user1);
-            _context.Add(user2);
-            _context.Add(user3);
-            _context.SaveChanges();
-
-            var category1 = new Category()
-            {
-                Name = "Best Ween Songs"
-            };
-
-            var category2 = new Category()
-            {
-                Name = "Most Underrated Ween Songs"
-            };
-
-            _context.Add(category1);
-            _context.Add(category2);
-            _context.SaveChanges();
-
-            var post1 = new Post()
-            {
-                Title = "Voodoo Lady",
-                Content = "One of the best songs on Chocolate & Cheese",
-                ImageLocation = "http://foo.gif",
-                CreateDateTime = DateTime.Now - TimeSpan.FromDays(10),
-                PublishDateTime = DateTime.Now - TimeSpan.FromDays(10),
-                IsApproved = true,
-                CategoryId = 1,
-                UserProfileId = 1
-            };
-
-            var post2 = new Post()
-            {
-                Title = "Ocean Man",
-                Content = "The song everyone knows Ween for",
-                ImageLocation = "http://foo.gif",
-                CreateDateTime = DateTime.Now - TimeSpan.FromDays(10),
-                PublishDateTime = DateTime.Now - TimeSpan.FromDays(10),
-                IsApproved = true,
-                CategoryId = 2,
-                UserProfileId = 1
-            };
-
-            var post3 = new Post()
-            {
-                Title = "Exactly Where I'm At",
-                Content = "First song from White Pepper. Starts the album off right.",
-                ImageLocation = "http://foo.gif",
-                CreateDateTime = DateTime.Now - TimeSpan.FromDays(10),
-                PublishDateTime = DateTime.Now - TimeSpan.FromDays(10),
-                IsApproved = true,
-                CategoryId = 2,
-                UserProfileId = 3
-            };
-
-            _context.Add(post1);
-            _context.Add(post2);
-            _context.Add(post3);
-            _context.SaveChanges();
-        }
+        }    
     }
 }
