+++ Tabloid-Fullstack/Repositories/IPostRepository.cs
@@ -10,10 +10,14 @@ namespace Tabloid_Fullstack.Repositories
         
         //GetByUserId by Sam Edwards
         List<PostSummary> GetByUserId(int userId);
+        
         Post GetById(int id);
+        
         List<ReactionCount> GetReactionCounts(int postId);
+        
         //Add by Sam Edwards
         void Add(Post post);
+        
         //Delete by Sam Edwards
         void Delete(int id);
     }
